Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  596a0ea22967f2716a7ed8e18a9292c94dbbbb3f
Make http_get test deterministic

It happened for QChris every other time that the server thread started
slower than the http_get tried to make the request. This left the http
server (and hence tox) hanging. So we now postpone the request until
the http server grabbed the port and only afterwards perform the
request. If the port is not grabbed within two seconds, we consider
the test failed. This makes the test deterministic.

Change-Id: Ie9611caa7d65d218b5f92ba1db682af6b6a7b095