Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  90f13c68f28a2632536c64d689613d332f1deb0e
Harmonize used Python 3 versions

The main tox.ini used Python 2.7, and 3.3.
The server's tox.ini used Python 2.7 and 3.4.
So they disagreed on the used Python 3.

We teach tox to use all three versions in both the main and the
server's tox configuration, as vanadium has 2.7, and 3.3 and stat1003
has 2.7 and 3.4.

Change-Id: I5c215292f13d2f247a48fd16e0d51de6770229c0