Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  f447c1426ae8304cb6815acefe147b7d9ec15e95
sitelinkview: Implemented "remove" toolbar with icon

This change moves the "remove" toolbar right next to the site id and displays an icon instead
of the button text. For now, the change applies locally to the sitelinkview only. The logic will
be reevaluated and moved when converting the statement interface.

Deployment note: Requires purging the parser cache.

Bug: T76212
Change-Id: I9dc1849b1f6eaa456def2c31771c6caa13e3cd72