Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  cbb4feb9b8bd5efefdbda60a5d23ef452e7de861
Move Repo Hooks classes to Hooks folder/namespace

The folder in Wikibase Client was called "hooks" (lowercase plural),
the folder in Wikibase Repo was called "Hook" (uppercase singular).
The two patches If466217 and If9295f0 make these two equal.

A lot more folders/namespaces need to be cleaned up the same way
but I want to do this in patches that are as small as possible.

Change-Id: If9295f0996ac2aa791de4169b97d76e2d5e0923a