Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  fe3e0bf37dc1dc827e566d54aa8d23506feba7ec
Add support for "safer" query wrapper

The "safer" query wrapper analyzes the queries it wraps and degrades them
if they would take too long to run.  We want that if we can get it.

DEPLOYMENT: You have to turn this on with a setting _and_ you'll need
I9560105c31d429df24d13232222740acba9b8e9a from the elasticsearch plugin
to be released and deployed before you turn it on.

Change-Id: I7c4fa048a04f1c98fa20e46805f166486ec260d3