Updated mediawiki/extensions Project: mediawiki/extensions/Translate  9daa315c47123162dd6bf144d3872553c32d5470
Make ElasticSearchTTMServer results consistent enough

Added secondary sort on _uid to enforce consistent order
of same-score results.

Implemented an algorithm which does second query to fetch
more source documents if the first query results many docs
with identical score. This should help in the cases the
translations are sparse, so we will return suggestions above
certain score by checking all documents above that score,
not just what we got in the first query. This also gives
more reliable "used X times" numbers in the UI.

Increased the default size from 50 to 100.

Here we assume that same score means that source text is
also the same, which is only guaranteed to be true if score
is 1. I use the score instead of the source text because it
makes the code simpler.

Bug: T76184
Change-Id: I269f849e6d398cbe140c41411a7b5eefd417c862