Updated mediawiki/extensions Project: mediawiki/extensions/SemanticForms  2af08263f06c8482da7a9b739f6cebc820cdcd48
mapping template fixes:combobox/tokens/radiobutton

Fixes mapping template to work with combobox and tokens input types.

Bug fix to remove error when form is viewed with editing disabled
when using mapping template.

Bug fix to remove warning when mapping template is used with radiobutton
input type.

With this fix, mapping template now works with dropdown, listbox,
radiobutton, checkboxes, combobox, and tokens input types and with
multiple instance templates.

There is one hack with a comment in the code at lines 1061-1065
of SF_FormPrinter.php to allow the tokens input type to work without the
need for the "list" parameter. Without this, the field is not correctly
tokenized when the form is submitted when it has multiple values in the
absence of the "list" parameter. Suggestions are appreciated for
alternative solutions to remove the hack.

Change-Id: Ia6e31581a07b08e380bbe3d27e90950e7a63f1f4