Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  77af3bb50c2c13c1af48d7d8448df3849af89fc0
Alpha experiment: Replace search bar with branding and icon

The input serves no purpose in JavaScript other
than to make the entire header area clickable and make it
possible to search from the page itself without JavaScript
That said, according to the Zero team people click the search
icon without a search term thinking it is a link.

I'd be interested in us using this space for other things.

Change-Id: I3bbea996d8de3d23bb768e9cc4ffe202ccca0ca1