Updated mediawiki/extensions Project: mediawiki/extensions/MassMessage  9b4c0c62731f4536c57411c14ce2fccf79c17907
Implement AJAX warnings for CreateMassMessageList

In Special:CreateMassMessageList, the title is checked to see if it
already exists using AJAX, and a warning is given if so.

The source is also checked using AJAX to make sure it exists and has
content model of wikitext or MassMessageListContent.

Finally, implements progressive disclosure for the Source textbox
in a way that does not break if Javascript is disabled.

Bug: T72523
Bug: T72471
Bug: T72520
Change-Id: I3f4bcd39eb937f2c5f51662267e60d2b5f4c48a1