Updated mediawiki/extensions Project: mediawiki/extensions/Math  8477a17a09268ba35987ea4ae62d59c6b3b7f29a
Fix for fatal caused by static call to MathRenderer::getError

MathRenderer::getError() uses $this, so it should not be called statically.
MathInputCheckTexvc should instead call its convertTexvcError method, which
constructs a MathRenderer object so it can call its getError method.

This code is pretty awful. But this way it won't throw fatal errors all the time.

Change-Id: Ic438b307a3b464651363b4cc16698c7d4320b253