Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  dc6572828c4a2c932feee031a811ad957638423a
Improved tree selection and item handling

It is now only possible to edit a list that has been selected in the
lists tree. Also, after adding a new list, it is automatically selected.

Change-Id: I91922c1a3d12b27a85fcc485c964d5e187a232b8

Usability improvements in Checklist dialogue

Window is now of reasonable size. Moved button for new lists down to the
other buttons.

Change-Id: Iabe4565d01fb4ac0dbe06a48f6c2ef1addcad7c8