Updated mediawiki/extensions Project: mediawiki/extensions/Flow  02f12dea4c809d78aa3de2a230604986b30e4505
Fix phpunit failure in SpamBlacklistTest::testSpam

This test could fail depending on the execution order of the tests.
Essentially it depended on initializing the spam whitelists and
blacklists, but if the blacklist had already been initialized
this was ignored.

This patch resets the blacklist where appropriate so state is not
retained between tests.

Bug: T70655
Change-Id: I2a5d7d6ac3794973a1c67734414d4ba98ee640e6