Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  34e42e1d736780643fd184d1328e36b4679182f5
Update SendBeaconReliability revid

Fixing the type to be an integer is not sufficient to solve the
problem.  Even with the type as string, the version is generated
equivalent to if it were an integer.

That meant changing it to an integer did not update the startup module
entry.

That in turn means the same URLs, which means Varnish continued
caching the wrong version.

Change-Id: I73d12f82ecc5be30a3d359eeec2da371c5f3ce30