Updated mediawiki/extensions Project: mediawiki/extensions/Flow  f35f2fea576d3af417b1e64f8bbc4604a4ed0126
TOC: Trigger the loadmore check after each load

The loadmore code was attempting to trigger the scroll.flow event,
but under some circumstances was triggering it against an unparented
target.  Our actual handler is not the target but an element in the
bubble path, so the unparented event did not trigger the right code.

This removes the unnecessary check in flow-board-loadmore.js, and
adjusts to trigger the 'scroll.flow' event on the correct elements.

Change-Id: I8ecb1c417e6a378aeb582002e669ec288cc8d076