Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  f4501ccd224f1d8dbc764e12f94db132e03ffb96
Only use mw.ustring when necessary

mw.ustring is really really slow. I've discovered that in a lot of modules
on enwiki, upwards of 2/3 of the total runtime gets used when mw.html
calls mw.ustring.gsub. This change checks whether any Unicode characters
are present, and if not, calls string.gsub instead.

Change-Id: Ia50061584be3901ae7428354c449236225c318db