Updated mediawiki/extensions Project: mediawiki/extensions/Cite  944b24542827f827cdfb8646f1c2e87aeb929b65
Do all of Cite's real work during unstrip

Cite has side effects that currently happen at the same time that the
preprocessor is running. This can cause oddities like references being out
of order, or references showing up in the reference list but never used in
the text. By moving the work to unstrip, only references actually used in
the final text will be present, and the numbering order will always match
the order they were used in the text. This depends on If83b0623 in core.

Change-Id: I0e136f952302b65f1abeeac5273bd842867e1dac