Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  c489e06543f074e5dbf6d2eccf907930a2e62ea6
Error reporting interface for alpha

This adds a 'Report an error' button at the bottom of the page.

Since talk pages are unlikely to be deployed to stable any time soon
on mobile, this provides a lightweight interface for providing 1-way
feedback about pages for editors who aren't ready to edit.

Change-Id: Ifd27f175fa9f0bcfcc8c286c834745a47b3c9991