Updated mediawiki/extensions Project: mediawiki/extensions/MultimediaViewer  7f4ce833d4974ab7b00d39a19f514bf91d0a8143
Log right-clicks on the image

Record how many users right-click on the image in MediaViewer, with the
assumption being they're intending to download the image. However, that
image has been resized and is not the original.

The event is logged even if the metadata panel is open, as the user probably
still intended to download.

Bug: T75962
Change-Id: I223ed957bcc60646adf9c9a00d2d9ca18ad128e6