Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  11cc324d0fdc7a9fa51db8a76f09254fe1a4f65f
Remove defaults for choices in test fixtures

The use of default values in test fixtures helped make the fixtures
more compact, but as the fixtures become more complex, it will
make them more difficult to read. This is the first step in removing
them. The next step will be to remove them for the setup property
of test fixtures (used by PHPUnit tests). Note that some tweaking on
the PHPUnit side was needed here. That'll be fleshed out in the next
step.

Also added a tidbit of fixture documentation (since comments are not
allowed in json).

Change-Id: I1e1e807deb86806f9c19e42dabeb857529e43e22