Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  43a38fc8270dc26a2739ea74936bfd74776e6ff2
If the target link was removed, don't show its card on the source link click

The page id will be still in cache, so the target card will be shown. So
do an extra check to see if the target link is still there.

We are not purging the page from the cache so that we can easily
get the link information in case translator want to undo this.

Bug: T74265
Change-Id: Ia835f8a1669091ddb250713aa09c9446ebe1a595