Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  1892942d6e0dbe65b84033093716ac07029d31e0
Check for localStorage availability using window.localStorage

Checking for boolean localStorage is not safe,
and checking for window.localStorage is the right thing.

Follow up to
I9c57c9d6a8dfc7bfa8f36336ef513af3a9254254

This doesn't fully address bug T86137, because it's possible
that there won't be any source languages for a given target language.
This wil be addressed in a subsequent patch.

Bug: T86137
Change-Id: I2084909e6aa5d1f5987bf0ac3399bed795a86a22