Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  0e6fba53799194b65ebdb37784163d1b3251d8fe
Fix lowercase namespace in GenericEventDispatcherTest

I did a regex search and this is the one and only namespace that
contains a lowercase word.

My plan is to move all tests in this folder to Wikibase\Repo\Tests.
But this will be done in later patches to keep this one easy to
review and conflict-free.

Change-Id: I1c1d2836c62279bdee8fbabf7b9cd06f344ca397