Updated mediawiki/extensions Project: mediawiki/extensions/MultimediaViewer  40738c832af93cb433f16ccc9b9029e6158d2d25
Display the file name in metadata panel

This patch adds the file name to the metadata panel, after the license
item. The namespace prefix is included and set to a static 'File:'.

The Title class already converts underscore to spaces.

The icon is from @Pginer-WMF.

PS: This has been a frequent annoyance for me, I'm glad to be working
on this patch!

Bug: T76680
Change-Id: I7d1f4ce67a6776ac017f8afe49cb3102b267af5c