Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  f187bd76ab81295c3ea03e82927e34ee6ffe75ed
Remove defaults for setup in test fixtures

This is the second step in removing defaults for test fixture data,
to make way for expanded fixture coverage (and the complexification
of fixtures that entails).

Defaults are not removed for legacy tests that do not use the shared
fixture data. Also, test fixture data setup is provided a few default
properties that are not relevant to the tests covered by the fixture
data. (TODOs add to placeholder fixture documentation for this.)

Finally, CentralNoticeTestFixutres is reorganized and PHPdoc'd a bit,
support for setting devices and buckets in setup is added, and
some property names are tweaked.

Legacy tests are left more or less alone.

Change-Id: I063a4b421b6e9c169121c8a0a855d9706fa4e0ec