Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  1fd5c4de88b9007b121ccaa17a7dee92db987cce
Do not assume that PHP_SAPI is cli when running jobs

It (probably) isn't on WMF

This passes an explicit parameter from where the method is called in the
job. This is a bit hacky since the parameter isn't in all the base
classes. Perhaps this could form a temporary solution to the immediate
error.

I should note, I believe this would work, however I have not tested
it with WMFs actual config...

Bug: T87040
Change-Id: I8f572dfa701cef2a77bba309f0e82c40e8629d22