Updated mediawiki/extensions Project: mediawiki/extensions/Math  688908675385a151dba78434264db574410f5920
Move integration tests to a new special page

The integration tests depend on the presence of external servers
and the user settings.
First, people are bothered by  failing test due to problems with
those servers and second, administrators need to run phpunit tests
to verify the configuration settings.
Therefore, we move the integration tests to a special page.

Bug: T87389
Change-Id: If66f13fe7fa16868cd4b1d8f0e0a6e3d49448b27