Updated mediawiki/extensions Project: mediawiki/extensions/Flow  2059e8187dfa9972f6ea6630dff5ccdb83b1ddcd
Call CachingObjectMapper::clear() from ObjectManager::clear()

The FlowUpdateRevisionContentLength maint script can occasionally
fail trying to insert a row that it should be updating.  It turns
out this is because the ObjectManager was cleared of stored data
and didn't know the row was loaded, while the CachingObjectMapper
still held a reference by primary key to the instantitated object.

Patch adds a new method ObjectMapper::clear() (which already existed
on CachingObjectMapper but was not being called) and calls it from
ObjectManager::clear() to ensure the rows are removed from both
places.

Bug: T86934
Change-Id: I1dfcc9d09ac79e6aeda7b6e7c6cd97b9e4edff1c