Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  ad795e85168a752f1c47504d7a38c66fbb0c26de
Drop dead code from snakview.variations.Value

See I0600d91. I decided to not do explicit null checks in favor of
simple comparisons (they are much more common in JS than they are in
PHP) because the data type is either a string or not. There is no
difference between it being null or false.

Change-Id: Id438672a407154f3018bd0bfdc1ea8af6dec57c1