Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  fbd41feaa6d85d157c99705c8835bd85d38bb591
Fixed errors on activation permission templates in PermissionManager

I had to change some of the logic on how the PermissionGridModel handles new values.

Before it always ignored the new value which was given to the set method and tried to figure out what value to use by checking the current group- and lockdown settings. This was needed because the grid always called the set method with a boolean parameter but we need to handle three different states. It worked fine, if only permissions were changed directly.

With permission templates, on the other hand, it caused problems because they call the set method with a numeric parameter which represents one of the triple state values. Because the model ignored the parameter given, it led to situations, where checkboxes in the grid got unchecked, if they were already checked at the time, a permission template got activated.

I resolved the problem by adding more parameter checks and an alternativ logic into the model for the case, that numeric parameters are given.

Although this resolved the problem with the unchecked checkboxes, it also added some more logic to the model which has an impact on the reaction time of the grid. With a fresh BlueSpice installation and a permission template with 3 rights it takes about 1 second to activate the template. For a template with all rights it takes about 30 seconds to do the same. Each additional right or namespace increases this time. To work arround this problem, we should maybe think about to limit the amount of possible rights in a template.

Change-Id: I280685488459e20571e42cb25e793c056620e338