Updated mediawiki/extensions Project: mediawiki/extensions/Flow  a3c10c692de719bdeeb9fc44ae5eba02115d7a7e
Update references on topic moderation change

When a topic changes to a moderation state >= hidden it should no longer
be included in categorization or whatlinkshere tables.  This is a bit of
an information leak and should be closed.

The patch notices when a topic transitions in visibility and re-runs the
reference recording code against all the current top revisions within the
topic.

Bug: T87793
Change-Id: I30e07f68e578bf9d92d6f443b28bbbfb29462f5c