Updated mediawiki/extensions Project: mediawiki/extensions/Translate  59d53e05c655994d88eb979e17ac3c63f4361dbb
Split language details to subpage on Special:SupportedLanguages

This is the only way I was able to make this page fast enough:
* We scan almost all of the page table
* We only pick up titles in translation namespaces
* We only pick up titles having subpage of requested language code
* We check all revisions of those pages and give one score for each
  rev_user_text for those

We could further limit the scans by only checking latest version of
each page.
However, given that editcount is a better measure of translation
activity than "edits on the tip of history", and that translation unit
pages usually do not have that long a history, I did not think that
necessary.

The language cloud is still output at the top of each page for easy
navigation.

$wgTranslateAuthorBlacklist is now applied to the list of translators.

Removed the html-output caching as the generation of the pages is now
fast enough: it is not generating tens of thousands of links, which
alone used to take multiple seconds.

The NS_PORTAL way of fetching translators is removed. Even though it was
faster on translatewiki.net, now the general method is fast enough and
gives good data.
Portal pages are not regularly updated anymore and gave stale data.
Issues were caused on wikis where NS_PORTAL namespace is defined
but is not used as in translatewiki.net.

Bug: T54728
Change-Id: I071b806a08d78e01bfe05bccfcbdb5e5d73220f4