Updated mediawiki/extensions Project: mediawiki/extensions/Flow  ac52cdaf52228ef84764128d8207258d45ec8688
Handle LQT moved thread stubs

These stubs contain only a redirect as the first reply and
are programatically generated by LQT when moving the thread.

This is a minimal implementation, rewriting the redirect into
a template so we can output something.   Prior to this patch
parsoid renders the '#REDIRECT ...' into a <link> element and
nothing is displayed.

Change-Id: I5883124f10e43bd6193f3e9d637efe827fc29e7b