Rewrite sort-gitmodules script in Python
The existing shell script used a multibyte character (U+20AC EURO SIGN)
as a delimiter for paste and tr. This caused the primary issue evident
in 5776dde57c43, which had to be reverted. The sorting was also locale-
dependent because the script did not override LC_COLLATE or LC_ALL.

Though these bugs are easy to fix, the script also assumed that each
section is exactly four lines long. I went ahead and rewrote it in
Python.

Change-Id: Ic4ce9ef9d5d179bfb98f0301e7a3d8e3e8463020