Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  87fc9196af20ff355e1b688dcfbbee2480147891
Add and remove <br> tags on Wikibase Repo special pages

They are not necessary because all elements do have outer margins.
But if looking at the page with no CSS the form elements are all in
a single line. The <br> tags are kind of a fail-safe for usability
(e.g. screenreaders). They don't have impact on how the pages look.

Change-Id: Ie1c3fb66f99b92044799838c8c86dc7ca991d85c