Updated mediawiki/extensions Project: mediawiki/extensions/Flow  a39a2185cfae497808da3b4ec6bfe9691f5d213e
Consider moderated posts to have 0 content length

The contentLength of a moderated topic or post will be considered 0.
This means a transition from unmoderated to moderated will show in
RC as a -, and restoring will show as a +. This includes a maintenance
script that will go back in time and update all existing revisions.

Bug; T89183
Change-Id: Ie0b17c71ac8c95ea166a56a665e0c9cc40c1bfc6