Updated mediawiki/extensions Project: mediawiki/extensions/Flow  bcf4163eb029ce80479acbcfd6b75f4234997146
Refactor import postprocessor interface

This interface was not very truthful, the commit action was calling a method
that indicated a full talkpage was imported, but really it was just a piece
of the page.

This patch removes the talkpage level methods, and move the call site of the
regular events to be post-commit.  Additionally documented the methods, and
switched the arguments to take the State objects.  This is a for a followup
patch which logs about imported topics.

Change-Id: I78fd0fdbd7c2d592e7287fa7f521c9a499385bb9