Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  3dfc2fd54f5d7a52747126e260023292d5631321
Promote TalkOverlay to stable

Talk pages will be visible for all experiences users[1] with the TalkOverlay (button in the footer).
Logged in users can contribute to a talk page.

Added browser tests to cover new functions in TalkOverlay.

Added JS workaround to add "Discussion" button on pages, where a cached version is delivered.

[1] Edit count > 5 or in beta/alpha

Bug: T54165
Change-Id: I1c938f79a164c30f33258dc729e83bb17eaaee6c