Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  c8727683c7a8dea9e43dc6308bb6579dcee11774
Column content style change: make it position: relative

This helps position: absolute elements in page content not jumping to
random positions in the screen

For example, in Bug T89445, there is a coordinate element going to
top left corner of page because of position: absolute. This brings it
back to the column content area.

This is also in consistent with .mw-body-content style of core.
It has position: relative style

Bug: T89445
Change-Id: I54802f8a7c2398f111ea7a7114c6441429bcf92f