Updated mediawiki/extensions Project: mediawiki/extensions/WikiEditor  6986f8dba8d6446ec66e841ab69ff7d35aa4cb34
Remove CSS definition for tables in dialogs

The only dialog with a table is the table creator dialog.
For this table a more specific definition exist in
jquery.wikiEditor.dialogs.config.css
These rules overwrite all here removed definitions except of
  wikiEditor-toolbar-dialog table { margin-top: 0.75em; }
Without this definition the margin-top slightly increase to the
definition from .wikitable with margin-top: 1em.

Change-Id: I0aab246db05cb4330287666f6a71b56d6b88b0af