Updated mediawiki/extensions Project: mediawiki/extensions/Flow  1ba3950bb22cefafdd22cb41375435c122ba2b0d
Give TOC some breathing room

The line of CSS I remove here was my previous attempt for this functionality,
which I had obviously misunderstood.

Part of this patch is also a minor fix (.top), removing an unused var, and
adding an @todo. All of them are somewhat related, but I can split it up in
multiple patches if anyone would prefer.

Bug: T85012
Change-Id: I5c6fd77a2625215920283346a721aefe1fedd9ab