Updated mediawiki/extensions Project: mediawiki/extensions/Gather  9e7c4bf76ec798720a5aec5705a1adc3dce5254c
Add warning to translators that this feature is experimental.

Per Jon, the product owner for this project, this patch adds warnings to the
translators that this feature is experimental and the messages may radically
change at any time without warning. That way they can decide whether they want
to expend the effort translating the strings.

Change-Id: Ie4f0d358148404f88a14b5cfd2c66713bac6dbcf