Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  d59d51733999ddfdfdece15f4c0093deca7f1d30
Warn if a TemplateData block already exists in related page

Check if there already is a TemplateData block in a related
page and warn the user if that is the case. Related pages are
either the /doc page or, if we are already in a /doc page, its
immediate parent.

Bug: T74062
Change-Id: I83212f14ecd2dcc18970d21ad0d833b461405390