Updated mediawiki/extensions Project: mediawiki/extensions/MultimediaViewer  3bb2b8185a45fbc0c01ad53f8810fdf9ed7ca4e7
Fix canvas height calculation

Canvas height is calculated as viewport height minus above-the-fold
height but the latter did not take into account the padding. This
was probably broken last autumn when the above-the-fold contents
were rearranged, but the max-height rule on the image masked the
error so we did not notice until that rule was dropped.

Bug: T89631
Change-Id: Id53cd9c176528da33b393e5ed807d6f2e0886413