Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  3cc6afaa750c01beb3675b1e9e4c6e04605b2923
Use wikitext instead of page prop when checking for existing TemplateData

On the edit page, we show the user a message if a related page
(parent of /doc page, or vice versa) already has templatedata.

This check was wrongly detecting that the parent page has templatedata
when editing '/doc'. The check shouldn't use the 'templatedata' API
since that also includes transcluded templatedata.

Look in the wikitext content instead.

Change-Id: I9788e2beb0809ac3e676a90cd0dfe519c005afc9