Updated mediawiki/extensions Project: mediawiki/extensions/AJAXPoll  fa506fd5302a9df9eb103aa09a4e497ad3ab3f51
updated AJAXPoll to work with MW 1.24

replaced .live() calls with .on()

wrapped all handlers in a function so they can be reapplied after the page is updated via the ajax call.

replaced sajax_do_call() that was generating a bunch of deprecation errors. I replaced it with the method outlined here:

https://www.mediawiki.org/wiki/ResourceLoader/Legacy_JavaScript

Bug: T86573

Change-Id: I44538be1a8e42bd53a2107e3724686ae184599e3