Updated mediawiki/extensions Project: mediawiki/extensions/NSFileRepo  a8e4f4fc6ab3059480950d6ab53318a2d8f282ae
Fix for invalid file upload

If a user selects a file that is not allowed by $wgFileExtensions and
tries to upload it (without noticing the message)

 $uploadForm->mDesiredDestName

is an empty string. In this case NSFileRepo causes a fatal error.

Jenkins breaks on this commit [1] because of the extensions dependency to "Lockdown". Therefor I'll do a manual verification.

[1] https://integration.wikimedia.org/ci/job/mwext-NSFileRepo-testextension-zend/3/console

Change-Id: Iada71324a867d9794dee3ce707236e2aca03132f