Updated mediawiki/extensions Project: mediawiki/extensions/Flow  4fac549c3b76bf883c160323cc853cc4d823458a
This is the initial re-activation of VE

Basically all the high-level stuff works:
* New topics
* New posts
* Replies
* Header
* Summaries

I did some refactoring, such as making an AbstractEditor and
using OOJS conventions (inheritance, their approach to static properties).

I came up with a convention for how editors should signal whether they use
preview, and whether the editor is currently empty.

Updated the VE loading and isSupported code.  There are probably some
simplifications the VE team can help us make at some point.

Fixed a problem where inputs were being 'expanded' (which was a no-op
for the wikitext editor, but causes VE to load in the new topic title
box).

The most obvious user-facing issues are:
* It's a bit slow on first load (not sure how evident this will be on
  real servers).  Some of this is specifically because of
  how we handle VE/wikitext conversion.
* It doesn't support MW-specific features, such as links(!)
* When you click the header pencil, it shows the wikitext first,
  until you focus the textarea

However, it should be reviewable as a first pass (it will not impact
users since it requires a config change to enable it).

Minor typo fix in flow-component-events.js

Bug: T88182
Change-Id: If273555cd8bb4633f06c70d4979196c31526cbf6