Updated mediawiki/extensions Project: mediawiki/extensions/Gather  9185cb186b1ca7ab9ee6316f71f2579e1b9c36b0
Add api action for listing collections of a user

Sandbox test url:
http://localhost:8080/w/index.php?title=Special:ApiSandbox&mobileaction=toggle_view_desktop#action=gather&format=json&owner=User
(Change User with the user you want to see.)

If user is not specified, it defaults to the logged in user.

If the owner specified does not match the logged in user the collections list
will only show the public ones.

Changes:
* Introduces ArraySerializable

Bug: T91440
Change-Id: I6092d444af229e1b34ef35686a07ffbb1d12d1c2