Updated mediawiki/extensions Project: mediawiki/extensions/OAuth  e0544b2f15129461a62af10b2f90e8775f6113e9
Provide a more specific error message on request token reuse

The default message, "something went wrong", is a little too vague when
the end user uses their browser's back button and re-clicks "Accept".
It's simple enough to save a placeholder under the key so we can
differentiate this situation and return a more specific error message.

Bug: T74654
Change-Id: I3161177ee24c9ad0b9603a410185704680feb2de