Updated mediawiki/extensions Project: mediawiki/extensions/ConfirmEdit  07b873532ea0ba46a889ff8441ccd6cf05076b44
Avoid parsing text for links in EditPageBeforeEditButtons hook

The code for the hook is passing in an empty new-text, so there's no
point in parsing that and the existing page text for links.

We pass false to shouldCheck's $isContent parameter (and an empty string
for $content) because getting a Content would break compatibility with
1.21, which we specifically preserved in I9529b7e8.

This also kills two entirely-unused parameters to the recently-added
SimpleCaptcha::editShowCaptcha().

Bug: T88661
Change-Id: Ie8a2d4da4c5bf0dcc8977172debbeb5913b56191