Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  7dd6e44901ad1a960b23b71d2cf1edd28e5c6c5d
Campaign: Invite editors to try CX while creating a new article

For the users who did not enable CX beta feature, show an invitation
dialog while landing in article creation page.

If the user dismiss that invitation, it won't be shown again.

If the user choose to try CX, the CX beta feature will be enabled
for the user and will take the user to Special:CX page.

This commit introduce the campaign concept, where a valid campaign
URL parameter act as a token to start using CX.

Also introducing configuration wgContentTranslationCampaigns.
It is an array of enabled campaigns. By default no campaigns are
enabled.

There will be tooling to measure which campaigns performing well
in follow up commits.

Bug: T89701
Change-Id: Ieac3cc5723f634a224d15c1b022e16db5b5ff040