Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  ffbc8e152047cd8e9b36a75c4f6aa33da96da116
TDG: Support optional text of 'deprecated' property

The deprecated property allows for either a boolean or text, so we
should allow the user to insert guidance text in case that property
is true. To achieve that, the model also defines 'textValue' for
boolean properties with text representation so an internal property can
be set to hold that value and inputs can be automatically built from the
property structure.

Bug: T90734
Change-Id: Iadc6abdcc0cf2721a311cf43847b306cb269b5e8