Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  996c1a227ef748fd4323f80ca20cf00d25fc8e24
QA: add a sleep to make Chrome happy. I don't like it.

Since we started using a modern version of Chrome, we have an issue
where Selenium clicks the "Show my notifications" link before
the javascript that triggers the overlay is in place.

Firefox seems to not have the issue, but Chrome does.

'sleep 1' fixes the problem in the smoke test build. We were unable
to find a better way, at least for now.

Change-Id: Id7fc83186d620cad73050315fd2808f9be5ffad5