Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  3820295f23844f18935259833c59eb9fec35c52f
Reorganize api calls and add sourceHandler and error message

Add a sourceHandler that deals with all source-related actions
like fetching template source and validating the existing
TemplateData string.

Also add a MessageDialog warning the user in case the TemplateData
is invalid. In this case, the user will be given the option either
to correct the JSON by hand or to start the editor to replace the
existing TemplateData with a new one.

Bug: T91730
Bug: T91325
Change-Id: I4e6d04f02565a02d8dcf5c70a575ab6433caa27f