Updated mediawiki/extensions Project: mediawiki/extensions/Thanks  ddf1310eea9c722c7596bab3ca0b2b0d2111a686
Flow container needs to be reset before each test

Pimple throws an exception when overriding something that was
already fetched from container.
This is currently causing some extensions trouble.

To prevent shared state between tests the flow container 
needs to be reset in the setUp() routine of all tests
accesing the container.

Bug: T91967
Change-Id: Ic80b137e6ca2f8ba6ad34ce0bdbb1319b81c0a6e