Updated mediawiki/extensions Project: mediawiki/extensions/Flow  fffbec8621d353d316d91785f92398832617b759
Hygiene: Use factory closures instead of Action classes

When this was first written the only option was specific classes,
but since then the wikidata team have updated the Action handling
to accept closures.  This patch removes the redundant classes that
only specify the explicit action name and replaces them with
Closures.

Change-Id: I0c9d96cee594975b32c9c22cf145e0f3c3749f7c