Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  27b9a3042af5d0c5d2a3602266eaac4be524b1a9
Dump the query when 'cirrusDumpQuery' is requested

Currently we dump the Cirrus query when 'cirrusDumpQuery=yes' is
requested, but requiring the parameter to be 'yes' is rather silly.

This change triggers the query dump when the parameter is requested,
regardless of its value. That means it can have any value, or no
value at all:

* ...&cirrusDumpQuery=yes
* ...&cirrusDumpQuery=true
* ...&cirrusDumpQuery=yep
* ...&cirrusDumpQuery=1
* ...&cirrusDumpQuery

Change-Id: I026c1ac28891a95de80764d5629ebcc3c9790a85