Updated mediawiki/extensions Project: mediawiki/extensions/Gather  9b6fd48e03016ec37d562cc45a3da70d5c72d025
Tweaks for watchlist overlay lit creation when keyboard is open

This frees up more space for the overlay when the keyboard is open and focuses
the input so the user can insert a list name.

Tweak the gutter of the watchstar at the top so that interface not so cramped.

Bug: T92251
Change-Id: I2cba2fe4f2a04248da17b6a1abf6fd0ad2fda519