Updated mediawiki/extensions Project: mediawiki/extensions/HierarchyBuilder  fab17a955e33c10741d3670061358e2607aad9d0
Changed the behavior of SelectFromHierarchy to match the other input
types.

Added a global config flag called HierarchyBuilder_LegacyMode to toggle
between the old behavior and the new behavior.

The new behavior returns the list of selected page names without
surrounding [[ and ]] notation. This is now the default behavior. Setting
the new global config flag mentioned above to true will change the
behavior to the original return format.

Note that this is not currently a breaking change. Saving the form will
simply update a page to use the configured behavior.

Change-Id: I4648d5ba5e7c2cea4021ee6b753d934073ab14b8