Updated mediawiki/extensions Project: mediawiki/extensions/Cite  530ef0ab9ccda7a5adef8547b2f5c644e156e636
Highlight target references

Highlight the target reference when the user clicks a reference link,
and also highlight the note in the text when he clicks the backlink.

This is done using the :target selector and a light blue background-color.
The color (and the rest of the code) are from English Wikipedia, from where
it has been copied to many other wikis, so the feature itself and especially
the color seem to be agreed on.

Browsers that don't support the :target selector (notably Internet Explorer 8
and below) will just ignore the new style and fall back to current behavior.
Wikis that have added this code to their common.css can remove it, but
of course nothing will break when they keep the duplicate code.

Bug: T10303
Change-Id: I5775daf6934516ce9c98841399e5a358195c1374